# docs/API/openapi_cn.yaml
openapi: 3.0.3
info:
  title: NogoChain Exchange API
  description: |
    # NogoChain 交易所 API 文档
    
    NogoChain 区块链节点的完整 HTTP API 接口文档，支持交易提交、查询、钱包管理、挖矿等功能。
    
    ## 特性
    
    - **高性能**: 支持高并发请求，内置速率限制
    - **安全性**: 支持 Admin Token 认证，结构化错误响应
    - **RESTful**: 遵循 REST 架构风格
    - **实时性**: WebSocket 支持实时事件订阅
    
    ## 基础 URL
    
    - **主网**: `http://main.nogochain.org:8080`
    - **本地开发**: `http://localhost:8080`
    
    ## 认证方式
    
    部分管理接口需要 Admin Token 认证，通过 `Authorization` 头传递：
    
    ```http
    Authorization: Bearer YOUR_ADMIN_TOKEN
    ```
    
    ## 响应格式
    
    所有接口返回 JSON 格式数据：
    
    ### 成功响应
    
    ```json
    {
      "data": {...},
      "requestId": "req_xxxxx"
    }
    ```
    
    ### 错误响应
    
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "错误描述",
        "details": {...},
        "requestId": "req_xxxxx"
      }
    }
    ```
    
    ## 速率限制
    
    API 实施速率限制以保护节点资源：
    
    - **默认限制**: 10 请求/秒，burst 20
    - **API Key**: 可申请提升限制（5 倍）
    - **限流头**: 响应中包含 `X-RateLimit-Limit`、`X-RateLimit-Remaining`、`X-RateLimit-Reset`
    
    ## 错误码
    
    错误码分为以下几类：
    
    - **1000-1999**: 验证错误
    - **2000-2999**: 资源未找到
    - **3000-3999**: 内部错误
    - **4000-4999**: 速率限制
    - **5000-5999**: 认证授权错误
    
    详细错误码请参考 [错误码参考文档](./错误码参考.md)
    
    ## SDK 和工具
    
    - **JavaScript SDK**: 即将发布
    - **Python SDK**: 即将发布
    - **Postman Collection**: 即将发布
    
    ## 支持
    
...
  version: 1.2.0
  contact:
...
servers:
...
paths:
...
  /tx:
    post:
      tags:
        - Transaction
      summary: 提交交易
      description: |
        向网络广播已签名的交易。
        
        交易会被添加到内存池，等待矿工打包。
      operationId: submitTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionSubmitRequest'
      responses:
        '200':
          description: 交易提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionSubmitResponse'
        '400':
          description: 交易无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 请求频率超限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /mining/submit:
    post:
      tags:
        - Mining
      summary: 提交工作量
      description: |
        提交挖矿工作量证明。
        
        矿工会验证并提交区块到网络。
      operationId: submitWork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWorkRequest'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitWorkResponse'
        '400':
          description: 无效的工作量
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /block:
    post:
      tags:
        - Mining
      summary: 提交区块
      description: |
        提交完整的区块（管理接口）。
        
        用于节点间区块同步。需要 Admin Token 认证。
      operationId: addBlock
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Block'
      responses:
        '200':
          description: 区块接受
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  reorged:
                    type: boolean
        '400':
          description: 区块无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 使用 ADMIN_TOKEN 作为 Bearer Token

  # ===========================================================================
  # Schemas
  # ===========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      description: 错误响应
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      properties:
        code:
          type: string
          description: 错误码
          example: VALIDATION_ERROR
        message:
          type: string
          description: 错误消息
          example: invalid request body
        details:
          type: object
          description: 详细错误信息
          additionalProperties: true
        requestId:
          type: string
          description: 请求 ID
          example: req_abc123

    # -------------------------------------------------------------------------
    # System
    # -------------------------------------------------------------------------
...
    Block:
      type: object
      properties:
        height:
          type: integer
        hash:
          type: string
        prevHash:
          type: string
        timestampUnix:
          type: integer
        difficultyBits:
          type: integer
        nonce:
          type: integer
        minerAddress:
          type: string
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...
    Transaction:
      type: object
      properties:
        type:
          type: string
          enum: [transfer, coinbase, contract, proposal]
        chainId:
          type: integer
        fromPubKey:
          type: string
          format: byte
        toAddress:
          type: string
        amount:
          type: integer
        fee:
          type: integer
        nonce:
          type: integer
        data:
          type: string
        signature:
          type: string
          format: byte
    TransactionSubmitRequest:
      type: object
      properties:
        rawTx:
          type: string
          description: 已签名的交易（十六进制）
      required:
        - rawTx
    TransactionSubmitResponse:
      type: object
      properties:
        accepted:
          type: boolean
        message:
          type: string
        txId:
          type: string
...
    SubmitWorkRequest:
      type: object
      properties:
        height:
          type: integer
        nonce:
          type: integer
        timestamp:
          type: integer
        minerAddress:
          type: string
      required:
        - height
        - nonce
        - timestamp
        - minerAddress
    SubmitWorkResponse:
      type: object
      properties:
        accepted:
          type: boolean
        message:
          type: string
        blockHash:
          type: string
        height:
          type: integer
...